DEMRUM-1571: move iOSdSYMUtils into dsyms directory
